requirements.txt
groq
streamlit
python-docx
